+++ backend/core/src/activity-log/decorators/activity-log-metadata.decorator.ts
@@ -1,0 +1,5 @@
+import { SetMetadata } from "@nestjs/common"
+import { ActivityLogMetadataType } from "../types/activity-log-metadata-type"
+
+export const ActivityLogMetadata = (metadata: ActivityLogMetadataType) =>
+  SetMetadata("activity_log_metadata", metadata)