+++ sites/public/cypress/integration/pages/submit-application.spec.ts
@@ -1,10 +1,12 @@
 import { coliseumApplication, minimalDataApplication } from "../../mockData/applicationData"
 
 describe("Submit", function () {
-  it("should submit an application for the Test: Coliseum listing", function (done) {
-    cy.submitApplication("Test: Coliseum", coliseumApplication, done)
+  it("should submit an application for the Test: Coliseum listing", function () {
+    cy.log("submit 1")
+    cy.submitApplication("Test: Coliseum", coliseumApplication)
   })
-  it("should submit a minimal application for the Test: Default, No Preferences", function (done) {
-    cy.submitApplication("Test: Default, No Preferences", minimalDataApplication, done, false)
+  it("should submit a minimal application for the Test: Default, No Preferences", function () {
+    cy.log("submit 2")
+    cy.submitApplication("Test: Default, No Preferences", minimalDataApplication, false)
   })
 })