+++ sites/public/pages/_app.tsx
@@ -9,14 +9,11 @@ import {
   AuthProvider,
   ConfigProvider,
   LoggedInUserIdleTimeout,
-  blankApplication,
 } from "@bloom-housing/ui-components"
+import { blankApplication } from "@bloom-housing/shared-helpers"
 import { headScript, bodyTopTag, pageChangeHandler } from "../src/customScripts"
 import { AppSubmissionContext } from "../lib/AppSubmissionContext"
-import ApplicationConductor, {
-  loadApplicationFromAutosave,
-  loadSavedListing,
-} from "../lib/ApplicationConductor"
+import ApplicationConductor, { loadSavedListing } from "../lib/ApplicationConductor"
 import { translations, overrideTranslations } from "../src/translations"
 import LinkComponent from "../src/LinkComponent"
 
@@ -24,7 +21,7 @@ function BloomApp({ Component, router, pageProps }: AppProps) {
   const { locale } = router
   //  const initialized = useState(true)
   const [application, setApplication] = useState(() => {
-    return loadApplicationFromAutosave() || blankApplication()
+    return { ...blankApplication }
   })
   const [savedListing, setSavedListing] = useState(() => {
     return loadSavedListing()
@@ -39,7 +36,7 @@ function BloomApp({ Component, router, pageProps }: AppProps) {
     if (locale && locale !== "en" && translations[locale]) {
       addTranslation(translations[locale])
     }
-
+    addTranslation(overrideTranslations.en)
     if (overrideTranslations[locale]) {
       addTranslation(overrideTranslations[locale])
     }