+++ sites/public/pages/applications/preferences/all.tsx
@@ -4,7 +4,6 @@ import {
   AlertBox,
   Form,
   FormCard,
-  ProgressNav,
   Field,
   t,
   ExtraField,
@@ -16,15 +15,15 @@ import {
   mapApiToPreferencesForm,
   getPreferenceOptionName,
   getExclusivePreferenceOptionName,
-  OnClientSide,
   getExclusiveKeys,
   setExclusive,
+  ProgressNav,
 } from "@bloom-housing/ui-components"
 import FormsLayout from "../../../layouts/forms"
 import FormBackLink from "../../../src/forms/applications/FormBackLink"
 import { useFormConductor } from "../../../lib/hooks"
 import { FormMetadataExtraData, Preference } from "@bloom-housing/backend-core/types"
-import { stateKeys } from "@bloom-housing/shared-helpers"
+import { stateKeys, OnClientSide } from "@bloom-housing/shared-helpers"
 
 const ApplicationPreferencesAll = () => {
   const clientLoaded = OnClientSide()
@@ -112,6 +111,7 @@ const ApplicationPreferencesAll = () => {
     }
     // Otherwise complete the section and move to the next URL
     conductor.completeSection(4)
+    conductor.sync()
     conductor.routeToNextOrReturnUrl()
   }
 
@@ -261,9 +261,9 @@ const ApplicationPreferencesAll = () => {
           currentPageSection={currentPageSection}
           completedSections={application.completedSections}
           labels={conductor.config.sections.map((label) => t(`t.${label}`))}
+          mounted={clientLoaded}
         />
       </FormCard>
-
       <FormCard>
         <FormBackLink
           url={conductor.determinePreviousUrl()}