+++ sites/public/pages/applications/preferences/general.tsx
@@ -8,13 +8,14 @@ import {
   AppearanceStyleType,
   Button,
   FormCard,
-  ProgressNav,
   t,
   Form,
+  ProgressNav,
 } from "@bloom-housing/ui-components"
 import FormsLayout from "../../../layouts/forms"
 import FormBackLink from "../../../src/forms/applications/FormBackLink"
 import { useFormConductor } from "../../../lib/hooks"
+import { OnClientSide } from "@bloom-housing/shared-helpers"
 
 const ApplicationPreferencesGeneral = () => {
   const [hideReviewButton, setHideReviewButton] = useState(false)
@@ -37,9 +38,9 @@ const ApplicationPreferencesGeneral = () => {
           currentPageSection={currentPageSection}
           completedSections={application.completedSections}
           labels={conductor.config.sections.map((label) => t(`t.${label}`))}
+          mounted={OnClientSide()}
         />
       </FormCard>
-
       <FormCard>
         <FormBackLink
           url={conductor.determinePreviousUrl()}