+++ ui-components/src/overlays/Modal.tsx
@@ -1,7 +1,6 @@
 import React from "react"
 import "./Modal.scss"
 import { Icon } from "../icons/Icon"
-import { GridCell, GridSection } from "../sections/GridSection"
 import { Overlay, OverlayProps } from "./Overlay"
 
 export interface ModalProps extends Omit<OverlayProps, "children"> {
@@ -19,12 +18,11 @@ const ModalHeader = (props: { title: string }) => (
 
 const ModalFooter = (props: { actions: React.ReactNode[] }) => (
   <footer className="modal__footer bg-primary-lighter" data-testid="footer">
-    <GridSection columns={4} reverse={true} tightSpacing={true}>
-      {props.actions &&
-        props.actions.map((action: React.ReactNode, index: number) => (
-          <GridCell key={index}>{action}</GridCell>
-        ))}
-    </GridSection>
+    <div className="flex flex-row-reverse gap-5">
+      {props.actions.map((action: React.ReactNode, index: number) => (
+        <div key={index}>{action}</div>
+      ))}
+    </div>
   </footer>
 )